move mock server into uabot_test.go
